Added football award edition
